backend/requirements.txt
# Core server
flask
flask-cors
gunicorn
pymongo
google-generativeai

# File parsing
pdfplumber
python-docx
pillow
pdf2image
lxml

# Utils
requests
python-dotenv
tqdm
google-api-python-client
